detailedpage shows right date converted from unix timestamp

+++ app/src/main/java/com/greenfox/fuchsit/zerdareader/model/NewsItem.java
@@ -3,9 +3,6 @@ package com.greenfox.fuchsit.zerdareader.model;
 import com.google.gson.annotations.SerializedName;
 
 import java.io.Serializable;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 
 /**
  * Created by Zsuzska on 2017. 01. 24..
@@ -19,7 +16,7 @@ public class NewsItem implements Serializable{
     @SerializedName("description")
     private String description;
     @SerializedName("created")
-    private int created;
+    private long created;
     @SerializedName("feed_name")
     private String feedName;
     @SerializedName("feed_id")
@@ -40,7 +37,7 @@ public class NewsItem implements Serializable{
         this.description = description;
     }
 
-    public NewsItem(String title, String description, int created, String feedName, long feedId, boolean favorite, boolean opened, String url) {
+    public NewsItem(String title, String description, long created, String feedName, long feedId, boolean favorite, boolean opened, String url) {
         this.title = title;
         this.description = description;
         this.created = created;
@@ -51,7 +48,7 @@ public class NewsItem implements Serializable{
         this.url = url;
     }
 
-    public NewsItem(String title, String description, int created, String feedName, boolean favorite, boolean opened) {
+    public NewsItem(String title, String description, long created, String feedName, boolean favorite, boolean opened) {
         this.title = title;
         this.description = description;
         this.created = created;
@@ -80,7 +77,7 @@ public class NewsItem implements Serializable{
         this.description = description;
     }
 
-    public int getCreated() {
+    public long getCreated() {
         return created;
     }
